+++ src/app/(authroutes)/layout.tsx
@@ -1,51 +1,31 @@
 import "../globals.css";
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import { ToggleProvider } from "@/contextProviders/ToggleProvider";
-import NextAuthProvider from "@/contextProviders/sessionProvider";
-import { Toaster } from "react-hot-toast";
+import Sidebar from "@/components/Sidebar";
+import MainContent from "@/components/MainContent";
 import { Providers } from "@/store/provider";
 import { Suspense } from "react";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "TPC Portal",
-  description: "Generated by create next app",
-};
 
 interface Props {
   children: React.ReactNode;
 }
 
-const AuthLayout = async ({ children }: Props) => {
-  const className = inter.className;
-
+const RouteLayout = async ({ children }: Props) => {
   return (
-    <html lang="en">
-      <body className={className}>
-        <ToggleProvider>
-          <div className="flex flex-col h-screen">
-            <Toaster />
-
-            <NextAuthProvider>
-              <div className="flex-auto flex h-[92vh] ">
-                {/* sidebar and main content share this space */}
-                <Suspense fallback={<>Loading...</>}></Suspense>
-                <Providers>
-                  <div className="w-screen min-h-screen h-max bg-[url(/images/iiti_bg.JPG)] bg-cover bg-fixed bg-center bg-no-repeat absolute">
-                    <div className="w-full min-h-full h-max bg-gradient-to-b from-black/40 to-black/90 bg-opacity-50 absolute">
-                      {children}
-                    </div>
-                  </div>
-                </Providers>
+    <div className="flex-auto flex h-[92vh] ">
+      <Providers>
+        <div className="flex-auto flex h-[92vh] ">
+          {/* sidebar and main content share this space */}
+          <Suspense fallback={<>Loading...</>}></Suspense>
+          <Providers>
+            <div className="w-screen min-h-screen h-max bg-[url(/images/iiti_bg.JPG)] bg-cover bg-fixed bg-center bg-no-repeat absolute">
+              <div className="w-full min-h-full h-max bg-gradient-to-b from-black/40 to-black/90 bg-opacity-50 absolute">
+                {children}
               </div>
-            </NextAuthProvider>
-          </div>
-        </ToggleProvider>
-      </body>
-    </html>
+            </div>
+          </Providers>
+        </div>
+      </Providers>
+    </div>
   );
 };
 
-export default AuthLayout;
+export default RouteLayout;
