+++ src/app/(routes)/layout.tsx
@@ -1,53 +1,24 @@
 import "../globals.css";
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import { ToggleProvider } from "@/contextProviders/ToggleProvider";
 import Sidebar from "@/components/Sidebar";
 import MainContent from "@/components/MainContent";
-import NextAuthProvider from "@/contextProviders/sessionProvider";
-import { Toaster } from "react-hot-toast";
 import { Providers } from "@/store/provider";
 import { Suspense } from "react";
 
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "TPC Portal",
-  description: "Generated by create next app",
-};
-
 interface Props {
   children: React.ReactNode;
-  auth: React.ReactNode;
 }
 
-const RouteLayout = async ({ children, auth }: Props) => {
-  const className = inter.className;
-
+const RouteLayout = async ({ children }: Props) => {
   return (
-    <html lang="en">
-      <body className={className}>
-        <ToggleProvider>
-          <div className="flex flex-col h-screen">
-            {/* Page Content */}
-            <Toaster />
-
-            <NextAuthProvider>
-              <div className="flex-auto flex h-[92vh] ">
-                {/* sidebar and main content share this space */}
-                <Suspense fallback={<>Loading...</>}>
-                  <Sidebar />
-                </Suspense>
-                <MainContent>
-                  <Providers>{children}</Providers>
-                </MainContent>
-              </div>
-              {auth}
-            </NextAuthProvider>
-          </div>
-        </ToggleProvider>
-      </body>
-    </html>
+    <div className="flex-auto flex h-[92vh] ">
+      {/* sidebar and main content share this space */}
+      <Suspense fallback={<>Loading...</>}>
+        <Sidebar />
+      </Suspense>
+      <MainContent>
+        <Providers>{children}</Providers>
+      </MainContent>
+    </div>
   );
 };
